Move TFJS implementation to a separate folder
