+++ handpose/src/tfjs/box.ts
