+++ handpose/src/tfjs/handpose_test.ts
@@ -19,7 +19,7 @@ import * as tf from '@tensorflow/tfjs-core';
 // tslint:disable-next-line: no-imports-from-dist
 import {ALL_ENVS, describeWithFlags} from '@tensorflow/tfjs-core/dist/jasmine_util';
 
-import * as handpose from './index';
+import * as handpose from '../index';
 import {stubbedImageVals} from './test_util';
 
 describeWithFlags('Handpose', ALL_ENVS, () => {
